+++ app/assets/i18n/pt.json
@@ -155,5 +155,17 @@
     "copy_run_text": {
         "success": "Texto copiado para a área de transferência com sucesso",
         "error": "Falha ao copiar o texto para a área de transferência"
+    },
+    "mark_favorite": {
+        "success": "Missão marcada como favorita com sucesso.",
+        "error": "Erro ao marcar a missão como favorita."
+    },
+    "remove_favorite": {
+        "success": "Missão removida dos favoritos com sucesso.",
+        "error": "Erro ao remover a missão dos favoritos."
+    },
+    "delete_run": {
+        "success": "Missão excluída com sucesso.",
+        "error": "Erro ao excluir a missão."
     }
 }
