+++ app/assets/i18n/ru.json
@@ -155,5 +155,17 @@
     "copy_run_text": {
         "success": "Текст успешно скопирован в буфер обмена",
         "error": "Не удалось скопировать текст в буфер обмена"
+    },
+    "mark_favorite": {
+        "success": "Миссия успешно добавлена в избранное.",
+        "error": "Ошибка при добавлении миссии в избранное."
+    },
+    "remove_favorite": {
+        "success": "Миссия успешно удалена из избранного.",
+        "error": "Ошибка при удалении миссии из избранного."
+    },
+    "delete_run": {
+        "success": "Миссия успешно удалена.",
+        "error": "Ошибка при удалении миссии."
     }
 }
