+++ app/lib/screens/home/widgets/run_analysis/overview/card_details.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+/// A data class to encapsulate card details.
+class CardDetails {
+  final Color color;
+  final IconData icon;
+  final String titleKey;
+  final double timeValue;
+
+  CardDetails(this.color, this.icon, this.titleKey, this.timeValue);
+}
